+++ info_linux.go
@@ -43,7 +43,7 @@ type SysInfo struct {
 	ForwardAckSegs  uint       `json:"fack segs"`    // # of forward ack'd segments in transmission queue
 }
 
-func info(s int) (*Info, error) {
+func info(s uintptr) (*Info, error) {
 	var sti sysTCPInfo
 	l := uint32(sizeofTCPInfo)
 	if err := getsockopt(s, ianaProtocolTCP, sysTCP_INFO, unsafe.Pointer(&sti), &l); err != nil {
