+++ sockopt_unix.go
@@ -12,40 +12,40 @@ import (
 	"time"
 )
 
-func setCork(s int, on bool) error {
+func setCork(s uintptr, on bool) error {
 	return setInt(s, &sockOpts[ssoCork], boolint(on))
 }
 
-func setKeepAliveIdleInterval(s int, d time.Duration) error {
+func setKeepAliveIdleInterval(s uintptr, d time.Duration) error {
 	d += (sockOpts[ssoKeepAliveIdleInterval].tmu - time.Nanosecond)
 	v := int(d / sockOpts[ssoKeepAliveIdleInterval].tmu)
 	return setInt(s, &sockOpts[ssoKeepAliveIdleInterval], v)
 }
 
-func setKeepAliveProbeInterval(s int, d time.Duration) error {
+func setKeepAliveProbeInterval(s uintptr, d time.Duration) error {
 	d += (sockOpts[ssoKeepAliveIdleInterval].tmu - time.Nanosecond)
 	v := int(d / sockOpts[ssoKeepAliveIdleInterval].tmu)
 	return setInt(s, &sockOpts[ssoKeepAliveProbeInterval], v)
 }
 
-func setKeepAliveProbeCount(s int, probes int) error {
+func setKeepAliveProbeCount(s uintptr, probes int) error {
 	return setInt(s, &sockOpts[ssoKeepAliveProbeCount], probes)
 }
 
-func getInt(s int, opt *sockOpt) (int, error) {
+func getInt(s uintptr, opt *sockOpt) (int, error) {
 	if opt.name < 1 || opt.typ != ssoTypeInt {
 		return 0, errOpNoSupport
 	}
-	v, err := syscall.GetsockoptInt(s, ianaProtocolTCP, opt.name)
+	v, err := syscall.GetsockoptInt(int(s), ianaProtocolTCP, opt.name)
 	if err != nil {
 		return 0, os.NewSyscallError("getsockopt", err)
 	}
 	return v, nil
 }
 
-func setInt(s int, opt *sockOpt, v int) error {
+func setInt(s uintptr, opt *sockOpt, v int) error {
 	if opt.name < 1 || opt.typ != ssoTypeInt {
 		return errOpNoSupport
 	}
-	return os.NewSyscallError("setsockopt", syscall.SetsockoptInt(s, ianaProtocolTCP, opt.name, v))
+	return os.NewSyscallError("setsockopt", syscall.SetsockoptInt(int(s), ianaProtocolTCP, opt.name, v))
 }
